Create Largest Rectangular Area in a Histogram
his approach is mainly an optimization over the previous approach.

When we compute next smaller element, we pop an item from the stack and mark current item as next smaller of it. One important observation here is the item below every item in stack is the previous smaller element. So we do not need to explicitly compute previous smaller. 

Below are the detailed steps of implementation.

Create an empty stack.

Start from the first bar, and do the following for every bar arr[i] where 'i' varies from 0 to n-1

If the stack is empty or arr[i] is higher than the bar at top of the stack, then push 'i' to stack. 
If this bar is smaller than the top of the stack, then keep removing the top of the stack while the top of the stack is greater. 

Let the removed bar be arr[tp]. Calculate the area of the rectangle with arr[tp] as the smallest bar. 

For arr[tp], the 'left index' is previous (previous to tp) item in stack and 'right index' is 'i' (current index).

If the stack is not empty, then one by one remove all bars from the stack and do step (2.2 and 2.3) for every removed bar
Largest-Rectangular-Area-in-a-Histogram-6.webp