Create Design a stack that supports getMin() in O(1) time
Using an Auxiliary Stack - O(1) Time and O(n) Space
Use two stacks: one to store actual stack elements and the other as an auxiliary stack to store minimum values. The idea is to do push() and pop() operations in such a way that the top of the auxiliary stack is always the minimum. Let us see how push() and pop() operations work.

Push(int x)

push x to the first stack (the stack with actual elements) 
compare x with the top element of the second stack (the auxiliary stack). Let the top element be y. 
If x is smaller than y then push x to the auxiliary stack. 
If x is greater than y then push y to the auxiliary stack.
int Pop()

pop the top element from the auxiliary stack. 
pop the top element from the actual stack and return it. Step 1 is necessary to make sure that the auxiliary stack is also updated for future operations.
int getMin()

Return the top element of the auxiliary stack.